refactor(main): colocate shell and home components
