+++ src/app/(main)/_components/EcommerceHero.tsx
