+++ src/ui/components/organisms/index.ts
@@ -10,14 +10,11 @@ export * from "./InvitationFormView";
 export * from "./CoupleInfoSection";
 export * from "./CreateGuestbookForm";
 export * from "./DeleteGuestbookForm";
-export * from "./EcommerceHero";
 export * from "./ErrorFallback";
 export * from "./FormField";
 export * from "./ImageField";
 export * from "./ImagePreviewItem";
 export * from "./ImagesSection";
-export * from "./LiveDemoSection";
-export * from "./MobileNav";
 export * from "./ParentsInfoSection";
 export * from "./PersonValueCard";
 export * from "./PaymentMethodSelector";
